Relocate some files
